Added working files to repo
