+++ emelia/classify_tickets.py
@@ -25,6 +25,7 @@ def main():
     df = pd.DataFrame(df, columns=['EventCause', 'DetectionMethod',
                                    'RestoreMethod', 'FixClassification',
                                    'Subsystem', 'Relevance'])
+    print(df)
 
     df['EventCause'] = pd.Categorical(df['EventCause'])
     df['DetectionMethod'] = pd.Categorical(df['DetectionMethod'])
